separates out GoogleSearchMap from SearchLocation page component
